@update/added icons to match real design
