+++ src/app/user-dashboard/layout.tsx
@@ -1,5 +1,3 @@
-
-
 "use client";
 
 import { useState } from "react";
